Move around the contents of product.rb, split out inner classes
